+++ tests/parser_spec.rs
@@ -141,6 +141,18 @@ fn should_parse_open_account_currency(#[case] input: &str, #[case] expected_curr
 }
 
 #[rstest]
+#[case("2014-05-01 open Assets:Checking", None)]
+#[case("2014-05-01 open Assets:Checking \"STRICT\"", Some("STRICT"))]
+#[case("2014-05-01 open Assets:Checking CHF \"STRICT\"", Some("STRICT"))]
+#[case("2014-05-01 open Assets:Checking \t \"STRICT\"", Some("STRICT"))]
+fn should_parse_open_account_booking_method(#[case] input: &str, #[case] expected: Option<&str>) {
+    let DirectiveContent::Open(open) = parse_single_directive(input).content else {
+        panic!("was not an open directive");
+    };
+    assert_eq!(open.booking_method.as_ref().map(|m| m.as_ref()), expected);
+}
+
+#[rstest]
 #[case("2014-05-01 close Assets:Cash", "Assets:Cash")]
 #[case("2014-05-01  close  Assets:Cash", "Assets:Cash")]
 #[case("2014-05-01\tclose\tAssets:Cash:Wallet", "Assets:Cash:Wallet")]